Move validations to separate file
